refactor(frontend): extract out pairing modal
