+++ frontend/src/routes/home/Home.scss
@@ -8,26 +8,6 @@
       flex-direction: row;
     }
   }
-
-  &__modal {
-    &-instruction {
-      margin-bottom: 1.5rem;
-      margin-top: 0.75rem;
-    }
-
-    &-partner {
-      align-items: center;
-      display: flex;
-
-      > * {
-        margin-right: 1rem;
-      }
-    }
-
-    &-partner-title {
-      margin-bottom: 1rem;
-    }
-  }
 }
 
 .practice-panel {
